+++ CHANGELOG.md
@@ -3,7 +3,14 @@ All notable changes to the "cpplint" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [0.0.6 - 0.0.7]
+
+### Add
+
+- add multi root support
+
 ## [0.0.5]
+
 - add line length configuration
 - add travis ci status
 - add exclude configuration
@@ -11,14 +18,18 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - add verbose configuration
 
 ## [0.0.4]
+
 - fix readme image
 - add source link
 
 ## [0.0.3]
+
 - add lintMode, able to choose scan single file or the whole workspace
 
 ## [0.0.2]
+
 - add missing dependence of lodash
 
 ## [0.0.1]
+
 - Initial release